move helpers methods to Service project
