+++ AvaloniaLogReader/App.axaml.cs
@@ -10,6 +10,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Services;
 using Services.Data;
+using Services.Helpers;
 using Services.Interfaces;
 using Services.Repositories;
 using Services.Stores;
