+++ Services/Helpers/FtpDownloader.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace WpfLogReader
+namespace Services.Helpers
 {
     public class FtpDownloader
     {
