+++ WpfLogReader/Commands/SettingsCommands/UnpackFilesCommand.cs
@@ -1,4 +1,5 @@
-﻿using Services.Stores;
+﻿using Services.Helpers;
+using Services.Stores;
 using System;
 using System.IO;
 using System.Threading.Tasks;
