+++ WpfLogReader/ViewModels/SettingsViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using Domain.Models;
+using Services.Helpers;
 using Services.Stores;
 using System;
 using System.Windows.Input;
